apps/ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg[binary]==3.1.19
SQLAlchemy==2.0.31
openai==1.37.1
python-dotenv==1.0.1

